# backend/src/main/java/com/himotech/laundryms/api/dto/request/CreateOrderRequest.java
package com.himotech.laundryms.api.dto.request;

import jakarta.validation.Valid;
import jakarta.validation.constraints.DecimalMin;
import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotNull;
import lombok.Data;

import java.math.BigDecimal;
import java.util.List;
import java.util.UUID;

@Data
public class CreateOrderRequest {
    private Long customerId;
    private CreateCustomerRequest customer;

    @NotNull(message = "createdByUserId is required")

    private UUID createdByUserId;
    private Integer serviceRateId;

    @NotNull(message = "weightKg is required")
    @DecimalMin(value = "0.01", message = "weightKg must be greater than 0")
    private BigDecimal weightKg;

    @Min(0)
    private Integer extraMinutes = 0;

    @Valid
    private List<AddOnInput> initialAddOns;
}